+++ src/main/java/janggi/view/ApplicationView.java
@@ -3,6 +3,11 @@ package janggi.view;
 import janggi.domain.Side;
 import janggi.domain.piece.Piece;
 import janggi.strategy.StrategyLabel;
+import janggi.view.input.Input;
+import janggi.view.output.Output;
+import janggi.view.resolver.PieceViewResolver;
+import janggi.view.resolver.SideViewResolver;
+import janggi.view.resolver.StrategyViewResolver;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.function.Supplier;
