+++ src/main/java/janggi/view/input/ConsoleReader.java
@@ -1,4 +1,4 @@
-package janggi.view;
+package janggi.view.input;
 
 import java.util.Arrays;
 import java.util.List;
