+++ src/main/java/janggi/view/resolver/PieceViewResolver.java
@@ -1,4 +1,4 @@
-package janggi.view;
+package janggi.view.resolver;
 
 import janggi.domain.Side;
 import janggi.domain.piece.Piece;
