+++ src/main/java/janggi/view/resolver/StrategyViewResolver.java
@@ -1,4 +1,4 @@
-package janggi.view;
+package janggi.view.resolver;
 
 import janggi.strategy.StrategyLabel;
